chore: organize webrtc code
