+++ lib/matrix.dart
@@ -29,13 +29,12 @@ export 'src/database/hive_collections_database.dart';
 export 'src/event.dart';
 export 'src/presence.dart';
 export 'src/event_status.dart';
-export 'src/voip/call.dart';
-export 'src/voip/group_call.dart';
+export 'src/voip/call_session.dart';
+export 'src/voip/group_call_session.dart';
 export 'src/voip/voip.dart';
-export 'src/voip/voip_content.dart';
-export 'src/voip/conn_tester.dart';
-export 'src/voip/utils.dart';
-export 'src/voip/voip_room_extension.dart';
+export 'src/voip/models/call_events.dart';
+export 'src/voip/utils/conn_tester.dart';
+export 'src/voip/utils/rtc_candidate_extension.dart';
 export 'src/room.dart';
 export 'src/timeline.dart';
 export 'src/user.dart';
