+++ lib/src/voip/utils/rtc_candidate_extension.dart
@@ -1,0 +1,9 @@
+import 'package:webrtc_interface/webrtc_interface.dart';
+
+extension RTCIceCandidateExt on RTCIceCandidate {
+  bool get isValid =>
+      sdpMLineIndex != null &&
+      sdpMid != null &&
+      candidate != null &&
+      candidate!.isNotEmpty;
+}
