+++ test/webrtc_stub.dart
@@ -3,6 +3,7 @@ import 'dart:typed_data';
 import 'package:webrtc_interface/webrtc_interface.dart';
 
 import 'package:matrix/matrix.dart';
+import 'package:matrix/src/voip/models/webrtc_delegate.dart';
 
 class MockWebRTCDelegate implements WebRTCDelegate {
   @override
@@ -27,7 +28,7 @@ class MockWebRTCDelegate implements WebRTCDelegate {
   }
 
   @override
-  Future<void> handleGroupCallEnded(GroupCall groupCall) async {
+  Future<void> handleGroupCallEnded(GroupCallSession groupCall) async {
     Logs().i('handleGroupCallEnded called in MockWebRTCDelegate');
   }
 
@@ -42,7 +43,7 @@ class MockWebRTCDelegate implements WebRTCDelegate {
   }
 
   @override
-  Future<void> handleNewGroupCall(GroupCall groupCall) async {
+  Future<void> handleNewGroupCall(GroupCallSession groupCall) async {
     Logs().i('handleNewGroupCall called in MockWebRTCDelegate');
   }
